refactor(trusty-mpm): split resume_error.rs to satisfy the SLOC gate
The #2577 review's WorkspaceGone/PaneGone enum split pushed lifecycle.rs
(1240 vs frozen budget 1237) and, after the split's new PaneGoneTmux
integration test, session_manager_mvp.rs (1521 vs 1500 cap) and
managed_routes/mod.rs (507 vs 500 cap) over their respective gates.

- Extracted ResumeManagedError + its From<ManagedError> impl + the
  unresumable_response 422 builder into a new managed_routes/resume_error.rs
  (mirrors the existing sibling-module pattern already used throughout
  managed_routes/). lifecycle.rs now just imports the type.
- Added impl IntoResponse for ResumeManagedError in resume_error.rs,
  centralising ALL status mapping (404/409/422x2/500) beside the variants —
  mod.rs's resume_managed_session handler collapses to a plain
  Ok/Err(e) => e.into_response() dispatch, comfortably clearing its cap
  instead of accumulating match-arm boilerplate.
- Extracted the ResumeManagedError-mapping tests + all four
  resume_managed_typed_* tests (NotFound/InvalidState/WorkspaceGone/PaneGone)
  plus the new PaneGoneTmux driver into a new sibling integration test
  tests/resume_unresumable_mapping.rs, mirroring the existing
  resume_reattach_tests.rs precedent for cap-pressured regression coverage.

No behavior change: cargo test -p trusty-mpm still reports 0 failed across
every suite; `bash scripts/check_line_cap.sh` reports 0 violations (no
allowlist bump).
